+++ backends/rapidpro/testdata.sql
@@ -37,7 +37,12 @@ INSERT INTO msgs_msg("id", "text", "high_priority", "created_on", "modified_on",
 INSERT INTO msgs_msg("id", "text", "high_priority", "created_on", "modified_on", "sent_on", "queued_on", "direction", "status", "visibility",
                         "msg_count", "error_count", "next_attempt", "external_id", "channel_id", "contact_id", "contact_urn_id", "org_id")
               VALUES(10001, 'test message without external', True, now(), now(), now(), now(), 'O', 'W', 'V',
-                     1, 0, now(), 'ext1', 10, 100, 1000, 1);
+                     1, 0, now(), '', 10, 100, 1000, 1);
+
+INSERT INTO msgs_msg("id", "text", "high_priority", "created_on", "modified_on", "sent_on", "queued_on", "direction", "status", "visibility",
+                        "msg_count", "error_count", "next_attempt", "external_id", "channel_id", "contact_id", "contact_urn_id", "org_id")
+              VALUES(10002, 'test message incoming', True, now(), now(), now(), now(), 'I', 'P', 'V',
+                     1, 0, now(), 'ext2', 10, 100, 1000, 1);
 
 /** Simple session */
 DELETE from flows_flowsession;