+++ handlers/test.go
@@ -59,6 +59,7 @@ type SendPrepFunc func(*httptest.Server, courier.ChannelHandler, courier.Channel
 type MockedRequest struct {
 	Method       string
 	Path         string
+	RawQuery     string
 	Body         string
 	BodyContains string
 }
@@ -78,6 +79,7 @@ type ChannelSendTestCase struct {
 	URNAuth              string
 	Attachments          []string
 	QuickReplies         []string
+	Topic                string
 	HighPriority         bool
 	ResponseToID         int64
 	ResponseToExternalID string
@@ -193,7 +195,7 @@ func RunChannelSendTestCases(t *testing.T, channel courier.Channel, handler cour
 		t.Run(testCase.Label, func(t *testing.T) {
 			require := require.New(t)
 
-			msg := mb.NewOutgoingMsg(channel, courier.NewMsgID(10), urns.URN(testCase.URN), testCase.Text, testCase.HighPriority, testCase.QuickReplies, testCase.ResponseToID, testCase.ResponseToExternalID)
+			msg := mb.NewOutgoingMsg(channel, courier.NewMsgID(10), urns.URN(testCase.URN), testCase.Text, testCase.HighPriority, testCase.QuickReplies, testCase.Topic, testCase.ResponseToID, testCase.ResponseToExternalID)
 
 			for _, a := range testCase.Attachments {
 				msg.WithAttachment(a)
@@ -218,7 +220,7 @@ func RunChannelSendTestCases(t *testing.T, channel courier.Channel, handler cour
 					require.Zero(testCase.ResponseBody, "ResponseBody should not be used when using testcase.Responses")
 					for mockRequest, mockResponse := range testCase.Responses {
 						bodyStr := string(body)[:]
-						if mockRequest.Method == r.Method && mockRequest.Path == r.URL.Path && (mockRequest.Body == bodyStr || (mockRequest.BodyContains != "" && strings.Contains(bodyStr, mockRequest.BodyContains))) {
+						if mockRequest.Method == r.Method && mockRequest.Path == r.URL.Path && mockRequest.RawQuery == r.URL.RawQuery && (mockRequest.Body == bodyStr || (mockRequest.BodyContains != "" && strings.Contains(bodyStr, mockRequest.BodyContains))) {
 							w.WriteHeader(mockResponse.Status)
 							w.Write([]byte(mockResponse.Body))
 							mockRRCount++